# tvet-timetable-generator/lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TVET Timetable Generator API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Class {
  id: number;
  name: string;
  trade: string;
  level: number;
}

export interface ClassInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  trade: string;
  level: number;
}

export interface Module {
  id: number;
  code: string;
  name: string;
  weeklyPeriods: number;
  isSpecific: boolean;
  trade: string;
  /** @nullable */
  teacherId?: number | null;
  /** @nullable */
  teacherName?: string | null;
  /** @nullable */
  teacherCode?: string | null;
}

export interface ModuleInput {
  /** @minLength 1 */
  code: string;
  /** @minLength 1 */
  name: string;
  /** @minimum 1 */
  weeklyPeriods: number;
  isSpecific: boolean;
  /** @minLength 1 */
  trade: string;
}

export interface Teacher {
  id: number;
  name: string;
  code: string;
  moduleCodes: string[];
}

export interface TeacherInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  code: string;
  moduleCodes: string[];
}

export interface FileUpload {
  filename?: string;
}

export interface UploadResult {
  imported: number;
  skipped: number;
  errors: string[];
  message?: string;
}

export interface GenerationResult {
  success: boolean;
  classesProcessed: number;
  entriesCreated: number;
  errors?: string[];
  message?: string;
}

export interface GenerationLog {
  id: number;
  createdAt: string;
  success: boolean;
  classesProcessed: number;
  entriesCreated: number;
  errors?: string[];
  /** @nullable */
  message?: string | null;
}

export interface DashboardStats {
  totalClasses: number;
  totalModules: number;
  totalTeachers: number;
  totalEntries: number;
  classesWithTimetables: number;
  classesWithoutTimetables: number;
  /** @nullable */
  lastGeneratedAt?: string | null;
}

export type TimetableCellType =
  (typeof TimetableCellType)[keyof typeof TimetableCellType];

export const TimetableCellType = {
  module: "module",
  cocurricular: "cocurricular",
  free: "free",
} as const;

export interface TimetableCell {
  period: number;
  day: number;
  type: TimetableCellType;
  /** @nullable */
  moduleName?: string | null;
  /** @nullable */
  moduleCode?: string | null;
  /** @nullable */
  teacherCode?: string | null;
  /** @nullable */
  teacherName?: string | null;
  /** @nullable */
  blockLabel?: string | null;
}

export interface TimetableGrid {
  classId: number;
  className: string;
  trade?: string;
  cells: TimetableCell[];
}

export interface TeacherTimetableCell {
  period: number;
  day: number;
  /** @nullable */
  className?: string | null;
  /** @nullable */
  moduleName?: string | null;
  /** @nullable */
  moduleCode?: string | null;
}

export interface TeacherTimetableGrid {
  teacherId: number;
  teacherName: string;
  teacherCode: string;
  cells: TeacherTimetableCell[];
}