# tvet-timetable-generator/lib/api-zod/src/generated/types/dashboardStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TVET Timetable Generator API
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardStats {
  totalClasses: number;
  totalModules: number;
  totalTeachers: number;
  totalEntries: number;
  classesWithTimetables: number;
  classesWithoutTimetables: number;
  /** @nullable */
  lastGeneratedAt?: string | null;
}